+++ src/data/models/configModel.ts
